requirements.txt
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
aiosqlite>=0.19.0
pandas>=2.0.0
ta>=0.11.0
pyyaml>=6.0
